Update password_generator.js
there's a bug in the generatePasswordSegment(includeNumbers) function. In the array of possible even numbers ([2,4,6,8]), its  not correctly selecting a random even number due to the use of Math.random() * 10. This can result in any digit from 0 to 9, not just the even numbers we want. We should change the range to be from 0 to 4 (exclusive), and then index into our array of even numbers.